+++ src/demos/DemoControl.vue
@@ -1,6 +1,6 @@
 <template>
 	<bulma-box>
-		<bulma-form-control :is-loading="load" :right-icon="rIcon" :tag="tag" left-icon="fas fa-envelope">
+		<bulma-form-control :is-loading="load" :left-icon="rIcon" :tag="tag" :is-expanded="load" :right-icon="lIcon">
 			<bulma-input placeholder="Email@domain.far" type="email" />
 
 		</bulma-form-control>
@@ -13,6 +13,8 @@
 	<bulma-form-field is-horizontal label="Loading">
 		<input v-model="load" type="checkbox" />
 	</bulma-form-field>
+
+
 </template>
 
 <script lang="ts" setup>
@@ -21,13 +23,16 @@
 	import BulmaInput from "../components/form/BulmaInput.vue";
 	import BulmaFormControl from "../components/form/groups/BulmaFormControl.vue";
 	import BulmaFormField from "../components/form/groups/BulmaFormField.vue";
-	import {BulmaStackedIcon} from "../types/IconTypes";
+	import {BulmaColouredIcon, BulmaMultiIcon} from "../types/IconTypes";
 
-	const expand = ref<boolean>(false);
 	const tag = ref<string>('div');
 	const load = ref<boolean>(false);
-	const rIcon = ref<BulmaStackedIcon>({
+	const rIcon = ref<BulmaMultiIcon>({
 		icon: ['fas fa-question fa-stack-1x', 'fa-regular fa-circle fa-stack-2x'],
 		stacked: 'fa-stack'
+	});
+	const lIcon = ref<BulmaColouredIcon>({
+		icon: 'fas fa-envelope',
+		colour: 'primary'
 	});
 </script>